airflow/requirements.txt
log4mongo==1.6.2
gevent==1.3.6
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
gspread==3.0.1
oauth2client==4.1.3
XlsxWriter==1.1.0
pymongo
beautifulsoup4
alembic
pandas

